Rewrite UserService, UserRoutes

+++ src/main/kotlin/org/nekrasov/routes/UserRoutes.kt
@@ -9,17 +9,29 @@ import org.nekrasov.domain.dto.request.UpdateUserDto
 import org.nekrasov.domain.dto.response.userToReadUserDto
 import org.nekrasov.domain.service.AuthService
 import org.nekrasov.domain.service.UserService
-import org.nekrasov.exceptions.IncompatibleQueryParameterTypeException
-import org.nekrasov.exceptions.MissingHeaderException
-import org.nekrasov.exceptions.MissingQueryParameterException
-import org.nekrasov.exceptions.UnauthorizedException
+import org.nekrasov.exceptions.*
 
 
 fun Route.userRoutes(authService: AuthService,
                      userService: UserService
 ){
-    route("/users") {
+    route("/api/v1/users") {
         get{
+            val token = call.request.headers["X-Auth-Token"] ?: throw MissingHeaderException("Missing  X-Auth-Token header")
+            if (!authService.checkToken(token))
+                throw UnauthorizedException("User not authorized")
+            val pageParameter = call.parameters["page"] ?: throw MissingQueryParameterException("Parameter page not specified in query")
+            val page = pageParameter.toLongOrNull() ?: throw IncompatibleQueryParameterTypeException("Parameter page requires the Long type")
+            val sizeParameter = call.parameters["size"] ?: throw MissingQueryParameterException("Parameter size not specified in query")
+            val size = sizeParameter.toIntOrNull() ?: throw IncompatibleQueryParameterTypeException("Parameter size requires the Int type")
+            if (size<=0 || page<=0)
+                throw IncompatibleQueryParameterTypeException("Parameter size and page more than zero")
+
+            val userList = userService.getAllUsers(page, size)
+            call.respond(HttpStatusCode.OK, userList.map(::userToReadUserDto))
+        }
+
+        get("/{id}"){
             val token = call.request.headers["X-Auth-Token"] ?: throw MissingHeaderException("Missing  X-Auth-Token header")
             if (!authService.checkToken(token))
                 throw UnauthorizedException("User not authorized")
@@ -31,24 +43,14 @@ fun Route.userRoutes(authService: AuthService,
             } ?: call.respond(HttpStatusCode.NotFound, mapOf("status" to "User not found"))
         }
 
-        get("chats"){
-            val token = call.request.headers["X-Auth-Token"] ?: throw MissingHeaderException("Missing  X-Auth-Token header")
-            if (!authService.checkToken(token))
-                throw UnauthorizedException("User not authorized")
-            val id = call.parameters["id"] ?: throw MissingQueryParameterException("Parameter id not specified in query")
-            val idUser = id.toLongOrNull() ?: throw IncompatibleQueryParameterTypeException("Parameter id requires the Long type")
-
-            userService.getUser(idUser)?.let {
-                call.respond(HttpStatusCode.OK, userService.getUserChats(idUser))
-            } ?: call.respond(HttpStatusCode.NotFound, mapOf("status" to "User not found"))
-        }
-
         patch{
             val token = call.request.headers["X-Auth-Token"] ?: throw MissingHeaderException("Missing  X-Auth-Token header")
             if (!authService.checkToken(token))
                 throw UnauthorizedException("User not authorized")
 
             val updateUserDto = call.receive<UpdateUserDto>()
+            if (!authService.checkUser(updateUserDto.id, token))
+                throw ForbiddenException("The current user does not have access to this information")
 
             if (userService.updateUser(updateUserDto)){
                 call.respond(HttpStatusCode.OK, mapOf("status" to "Ok"))
@@ -57,12 +59,15 @@ fun Route.userRoutes(authService: AuthService,
             }
         }
 
-        delete {
+        delete("/{id}") {
             val token = call.request.headers["X-Auth-Token"] ?: throw MissingHeaderException("Missing  X-Auth-Token header")
             if (!authService.checkToken(token))
                 throw UnauthorizedException("User not authorized")
             val id = call.parameters["id"] ?: throw MissingQueryParameterException("Parameter id not specified in query")
             val idUser = id.toLongOrNull() ?: throw IncompatibleQueryParameterTypeException("Parameter id requires the Long type")
+
+            if (!authService.checkUser(idUser, token))
+                throw ForbiddenException("The current user does not have access to this information")
 
             if (userService.deleteUser(idUser)){
                 call.respond(HttpStatusCode.OK, mapOf("status" to "Ok"))
@@ -71,26 +76,19 @@ fun Route.userRoutes(authService: AuthService,
             }
         }
 
-        get("/all"){
+        get("/{id}/chats"){
             val token = call.request.headers["X-Auth-Token"] ?: throw MissingHeaderException("Missing  X-Auth-Token header")
             if (!authService.checkToken(token))
                 throw UnauthorizedException("User not authorized")
+            val id = call.parameters["id"] ?: throw MissingQueryParameterException("Parameter id not specified in query")
+            val idUser = id.toLongOrNull() ?: throw IncompatibleQueryParameterTypeException("Parameter id requires the Long type")
 
-            val userList = userService.getAllUsers()
-            call.respond(HttpStatusCode.OK, userList.map(::userToReadUserDto))
+            if (!authService.checkUser(idUser, token))
+                throw ForbiddenException("The current user does not have access to this information")
+
+            userService.getUser(idUser)?.let {
+                call.respond(HttpStatusCode.OK, userService.getUserChats(idUser))
+            } ?: call.respond(HttpStatusCode.NotFound, mapOf("status" to "User not found"))
         }
-        get("/all_pag"){
-            val token = call.request.headers["X-Auth-Token"] ?: throw MissingHeaderException("Missing  X-Auth-Token header")
-            if (!authService.checkToken(token))
-                throw UnauthorizedException("User not authorized")
-            val pageParameter = call.parameters["page"] ?: throw MissingQueryParameterException("Parameter page not specified in query")
-            val page = pageParameter.toLongOrNull() ?: throw IncompatibleQueryParameterTypeException("Parameter page requires the Long type")
-            val sizeParameter = call.parameters["size"] ?: throw MissingQueryParameterException("Parameter size not specified in query")
-            val size = sizeParameter.toIntOrNull() ?: throw IncompatibleQueryParameterTypeException("Parameter size requires the Int type")
-
-            val userList = userService.getAllUsersPaginated(page, size)
-            call.respond(HttpStatusCode.OK, userList.map(::userToReadUserDto))
-        }
-
     }
 }
